+++ src/test/kotlin/tornadofx/testapps/InlineFragmentTest.kt
@@ -10,7 +10,7 @@ class InlineFragmentTestApp : App(DangerButtonView::class)
 
 class DangerButtonView : View("Do not click the button!") {
     override val root = stackpane {
-        setPrefSize(400.0, 100.0)
+        setPrefSize(400.0, 200.0)
         button("Don't click me") {
             style {
                 fontSize = 20.px
@@ -18,7 +18,7 @@ class DangerButtonView : View("Do not click the button!") {
                 textFill = Color.RED
             }
             setOnAction {
-                adhocWindow("What do you want?", stageStyle = UNDECORATED) {
+                openInternalAdhocWindow("What do you want?", overlayPaint = c(Color.RED.toString(), 0.4)) {
                     vbox(10) {
                         style {
                             padding = box(20.px)