+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [1.5.10-SNAPSHOT]
-
+- InternalWindow now has `overlayPaint` that defaults `c("#000", 0.4)
+- openInternalAdHocWindow added
 - `ObservableList.asyncItems` and `ListProperty.asyncItems`
 - `confirm()` function that executes an action if the user confirms
 - di delegate overload to support injecting a dependency by name (in addition to type)